services/backend/requirements.txt
# prod
asyncpg==0.27.0
bcrypt==4.0.1
fastapi==0.88.0
passlib==1.7.4
pydantic[dotenv]
yara-python
uvicorn==0.20.0

# dev
pytest==7.2.0
httpx==0.23.1